# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IUser": {
				"properties": {
					"id": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"first_name",
					"last_name",
					"email",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAccessToken": {
				"properties": {
					"access_token": {
						"type": "string"
					}
				},
				"required": [
					"access_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRefrshToken": {
				"properties": {
					"refresh_token": {
						"type": "string"
					}
				},
				"required": [
					"refresh_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TTokens": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IAccessToken"
					},
					{
						"$ref": "#/components/schemas/IRefrshToken"
					}
				]
			},
			"IAuthUser": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/IUser"
					},
					"authentication": {
						"$ref": "#/components/schemas/TTokens"
					}
				},
				"required": [
					"user",
					"authentication"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBadResponse": {
				"properties": {
					"reason": {
						"type": "string"
					}
				},
				"required": [
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISignUp": {
				"properties": {
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"pws": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"first_name",
					"last_name",
					"email",
					"pws",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISignIn": {
				"properties": {
					"email": {
						"type": "string"
					},
					"pws": {
						"type": "string"
					},
					"remember_me": {
						"type": "boolean"
					}
				},
				"required": [
					"email",
					"pws",
					"remember_me"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPhotographer": {
				"properties": {
					"photograph_avatar": {
						"type": "string"
					},
					"full_name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"average_raiting": {
						"type": "number",
						"format": "double"
					},
					"count_of_reviews": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"photograph_avatar",
					"full_name",
					"description",
					"average_raiting",
					"count_of_reviews"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPhotographerParams": {
				"properties": {
					"price_per_hour": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"types_of_photography": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "canie-backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Val16eria ",
			"email": "KStorer61@icloud.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/signup": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IAuthUser"
										},
										{
											"$ref": "#/components/schemas/IBadResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"user": {
												"id": "1",
												"first_name": "Valeria",
												"last_name": "Mikhailenko",
												"email": "test@gmail.com",
												"role": "photograph"
											},
											"authentication": {
												"access_token": "your access_token",
												"refresh_token": "your refresh_token"
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBadResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignUp"
							}
						}
					}
				}
			}
		},
		"/auth/signin": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IAuthUser"
										},
										{
											"$ref": "#/components/schemas/IBadResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"user": {
												"id": "1",
												"first_name": "Valeria",
												"last_name": "Mikhailenko",
												"email": "test@gmail.com",
												"role": "photograph"
											},
											"authentication": {
												"access_token": "your access_token",
												"refresh_token": "your refresh_token"
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBadResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignIn"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/IBadResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": "Logout successful"
									}
								}
							}
						}
					},
					"400": {
						"description": "Logout Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBadResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			}
		},
		"/photographers": {
			"get": {
				"operationId": "PhotographersList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IPhotographer"
											},
											"type": "array"
										},
										{
											"$ref": "#/components/schemas/IBadResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"photograph_avatar": "https://i.pinimg.com/736x/ff/75/e7/ff75e70833921cbb8ff7e2c4d36167bd.jpg",
												"full_name": "Valeria Mikhailenko",
												"description": "For more than 3 years I have been specializing in shooting urban and natural objects, as well as portraits",
												"average_raiting": 5,
												"count_of_reviews": 345
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Photographer Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBadResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Photographers"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "price_per_hour",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "number",
								"format": "double"
							}
						}
					},
					{
						"in": "query",
						"name": "types_of_photography",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/token": {
			"post": {
				"operationId": "Token",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IAccessToken"
										},
										{
											"$ref": "#/components/schemas/IBadResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"access_token": "access_token"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Token Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBadResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Token"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRefrshToken"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}